+++ src/Easter.cpp
@@ -17,7 +17,14 @@ Easter::Easter() {
 
 void Easter::process(const ProcessArgs &args) {
 
-	if (outputs[OUTPUT_MAIN].isConnected()) {
+	if (upTrigger.process(rescale(params[PARAM_UP].getValue(), 1.f, 0.1f, 0.f, 1.f))) 
+		afp.algorithm = filterAlgorithm::kResonB;	
+	if (downTrigger.process(rescale(params[PARAM_DOWN].getValue(), 1.f, 0.1f, 0.f, 1.f)))
+		afp.algorithm = filterAlgorithm::kResonA;
+	afp.strAlgorithm = audioFilter.filterAlgorithmTxt[static_cast<int>(afp.algorithm)];
+	audioFilter.setParameters(afp);
+
+	if (outputs[OUTPUT_MAIN].isConnected() && inputs[INPUT_MAIN].isConnected()) {
 		audioFilter.setSampleRate(args.sampleRate);
 	
 		float cvfc = 1.f;
@@ -32,12 +39,6 @@ void Easter::process(const ProcessArgs &args) {
 		afp.Q = params[PARAM_Q].getValue() * rescale(cvq,-10,10,0,1);
 		afp.dry = params[PARAM_DRY].getValue();
 		afp.wet = params[PARAM_WET].getValue();
-	
-		if (upTrigger.process(rescale(params[PARAM_UP].getValue(), 1.f, 0.1f, 0.f, 1.f))) 
-			afp.algorithm = filterAlgorithm::kResonB;	
-		if (downTrigger.process(rescale(params[PARAM_DOWN].getValue(), 1.f, 0.1f, 0.f, 1.f))) {
-		afp.algorithm = filterAlgorithm::kResonA;
-		}
 
 		afp.strAlgorithm = audioFilter.filterAlgorithmTxt[static_cast<int>(afp.algorithm)];
 		audioFilter.setParameters(afp);
@@ -72,12 +73,17 @@ struct EasterModuleWidget : ModuleWidget {
 		addChild(createWidget<ScrewSilver>(Vec(0, 0)));
 		addChild(createWidget<ScrewSilver>(Vec(box.size.x - 15, 365)));
 
-		box.size = Vec(6*RACK_GRID_WIDTH, RACK_GRID_HEIGHT);
+		box.size = Vec(MODULE_WIDTH*RACK_GRID_WIDTH, RACK_GRID_HEIGHT);
 
 		{
-			ATitle * title = new ATitle(box.size.x);
+			RPJTitle * title = new RPJTitle(box.size.x,MODULE_WIDTH);
 			title->setText("EASTER");
 			addChild(title);
+		} 
+		{
+			RPJTextLabel * tl = new RPJTextLabel(Vec(1, 19),10,MODULE_WIDTH);
+			tl->setText("Resonance Filter");
+			addChild(tl);
 		}
 		{
 			FilterNameDisplay * fnd = new FilterNameDisplay(Vec(39,30));