+++ src/Easter.hpp
@@ -2,6 +2,8 @@
 #include "AudioFilter.hpp"
 
 using namespace rack;
+
+const int MODULE_WIDTH=6;
 		
 struct Easter : Module {